requirements.txt
APScheduler==3.10.4
Django==4.1
django_apscheduler==0.6.2
djangorestframework==3.14.0
IPy==1.1
python_nmap==0.7.1
mysqlclient==2.2.0
xlrd==2.0.1
xlwt==1.3.0
asgiref==3.5.2
Django==4.1
djangorestframework==3.14.0
kafka-python==2.0.2
mysqlclient==2.2.0
sqlparse==0.4.4
psutil==5.9.6
